requirements.txt
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
dimod==0.9.1
docutils==0.16
dwave-cloud-client==0.7.0
dwave-hybrid==0.4.2
dwave-inspector==0.1.4
dwave-inspectorapp==0.1.4
dwave-neal==0.5.3
dwave-networkx==0.8.5
dwave-ocean-sdk==2.0.1
dwave-qbsolv==0.3.0
dwave-system==0.9.1
dwave-tabu==0.2.2
dwavebinarycsp==0.1.2
Flask==1.1.1
homebase==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
minorminer==0.1.9
networkx==2.4
numpy==1.18.2
ortools==7.5.7466
pandas==1.0.3
penaltymodel==0.16.3
penaltymodel-cache==0.4.1
penaltymodel-lp==0.1.3
penaltymodel-mip==0.2.2
plotly==4.5.4
plucky==0.4.3
protobuf==3.11.3
pycodestyle==2.5.0
pyqubo==0.4.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
